Refactor to use useClipboard() hook
